refactor: split `BaseGenerator` and `FewShotGenerator` and fix bugs in `FewShotGenerator`
